tests/test_requirements.txt
# Testing dependencies for Dukira Webhook API

# Core testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0

# HTTP testing
httpx==0.25.2
requests-mock==1.11.0

# Database testing
pytest-postgresql==5.0.0
sqlalchemy-utils==0.41.1

# Async testing support
asyncio-pytest==0.21.1

# Coverage reporting
coverage==7.3.2

# Test utilities
factory-boy==3.3.0
faker==19.13.0

# Linting and formatting (for test quality)
flake8==6.1.0
black==23.11.0
isort==5.12.0